moved attendee file to respective folders
